Consolidate FIN storage into recordings.db (PR #123 review)
Addresses review feedback on PR #123: drop the separate data/phi.db and store FIN alongside other recording metadata in the existing recordings.db. Proper PHI access control is deferred to the DataJoint export layer (future scope, @ktshah04).

- Delete multi_camera/backend/phi_db.py.
- Add ParticipantFIN model to recording_db.py with an integer FK to participants.id (matches the convention used by Session/Recording/Imported/Photo). Unique on participant_id (one FIN per participant, upserted on conflict).
- Add fin_record back-reference on the Participant model.
- Factor out _get_or_create_participant() from _get_or_create_session() so store_fin() can lazily create the participant row when (and only when) a FIN is provided. The no-FIN session-creation path is unchanged and still does not write to the DB.
- Move store_fin() and get_fin() helpers into recording_db.py.
- fastapi.py: drop phi_db_dependency, switch set_session() to the existing db_dependency, update the docstring, point store_fin import at recording_db. The API contract is unchanged (POST /api/v1/session?subject_id=...&fin=...).
- Update the push_to_datajoint() TODO to reference get_fin() in the same module and note that proper PHI access control belongs in the export layer.
