+++ multi_camera/backend/fastapi.py
@@ -107,15 +107,6 @@ def db_dependency():
         db.close()
 
 
-def phi_db_dependency():
-    from multi_camera.backend.phi_db import get_phi_db
-    db = get_phi_db()
-    try:
-        yield db
-    finally:
-        db.close()
-
-
 class ConnectionManager:
     def __init__(self):
         self.active_connections: list[WebSocket] = []
@@ -334,14 +325,15 @@ async def stop_recording():
 
 
 @api_router.post("/session", response_model=Session)
-async def set_session(subject_id: str, fin: Optional[str] = None, phi_db=Depends(phi_db_dependency)) -> Session:
+async def set_session(subject_id: str, fin: Optional[str] = None, db=Depends(db_dependency)) -> Session:
     """
     Create a new session directory for the participant
 
     Args:
         subject_id (str): The participant ID
         fin (str, optional): Financial Identification Number from patient wristband.
-            Stored in a separate PHI database (data/phi.db) with restricted access.
+            Stored in the recordings.db alongside the participant. Proper PHI access
+            control is deferred to the DataJoint export layer (future work).
     Returns:
         dict: A dictionary with the recording directory and filename
     """
@@ -357,8 +349,8 @@ async def set_session(subject_id: str, fin: Optional[str] = None, phi_db=Depends
     print("New session: ", state.current_session)
 
     if fin and fin.strip():
-        from multi_camera.backend.phi_db import store_fin
-        store_fin(phi_db, participant_name=subject_id, fin=fin.strip())
+        from multi_camera.backend.recording_db import store_fin
+        store_fin(db, participant_name=subject_id, fin=fin.strip())
         print(f"FIN stored for participant {subject_id}")
 
     return state.current_session
